Moved model, controller and pipes to own folders
