+++ src/app/app.module.ts
@@ -11,9 +11,9 @@ import { PlayerComponent } from './player/player.component';
 import { SongListComponent } from './song-list/song-list.component';
 import { AppRoutingModule } from './app-routing.module';
 import { MatTableModule, MatFormFieldModule, MatInputModule, MatProgressSpinnerModule, MatIconModule } from '@angular/material';
-import { MinutesSecondsPipe } from './minutes-seconds.pipe';
+import { MinutesSecondsPipe } from './pipes/minutes-seconds.pipe';
 import { SongDetailsComponent } from './song-details/song-details.component';
-import { FileIdToFilePipe } from './file-id-to-file.pipe';
+import { FileIdToFilePipe } from './pipes/file-id-to-file.pipe';
 import { MaterialModule } from './material.module'; 
 
 @NgModule({
