+++ src/app/services/song.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { catchError, map, tap } from 'rxjs/operators';
-import { Song } from './song';
+import { Song } from '../model/song';
 import { MessageService } from './message.service';
-import { environment } from './../environments/environment';
+import { environment } from '../../environments/environment';
 
 
 @Injectable({
